# swagger.json
{"openapi":"3.1.0","servers":[{"url":"http://localhost:3000","description":"Local Server"}],"info":{"title":"File Upload API","version":"1.0.0","description":"File Upload Service API Documentation","license":{"name":"MIT"}},"paths":{"/":{"get":{"tags":["helloworld"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"x-samchon-accessor":["getHello"],"x-samchon-controller":"AppController"}},"/healthcheck":{"get":{"tags":["helloworld"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheckResult"}}}}},"x-samchon-accessor":["healthcheck","healthCheck"],"x-samchon-controller":"AppController"}},"/files/upload":{"post":{"tags":["Files"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"storageType":{"oneOf":[{"const":"local"},{"const":"s3"}]}},"required":[]}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]}}}}},"x-samchon-accessor":["files","upload","uploadFile"],"x-samchon-controller":"FilesController"}},"/files/{id}/info":{"get":{"tags":["Files"],"parameters":[{"name":"id","in":"path","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileDto"}}}}},"x-samchon-accessor":["files","info","getFileInfo"],"x-samchon-controller":"FilesController"}},"/files/{id}/download":{"get":{"tags":["Files"],"parameters":[{"name":"id","in":"path","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{}}}},"x-samchon-accessor":["files","download","downloadFile"],"x-samchon-controller":"FilesController"}},"/users":{"post":{"tags":["Users"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegisterDto"}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{}}}}},"x-samchon-accessor":["users","createUser"],"x-samchon-controller":"UsersController"},"get":{"tags":["Users"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}},"x-samchon-accessor":["users","findUsers"],"x-samchon-controller":"UsersController"}},"/users/{id}":{"put":{"tags":["Users"],"parameters":[{"name":"id","in":"path","schema":{"type":"string"},"required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{}}}}},"x-samchon-accessor":["users","updateUser"],"x-samchon-controller":"UsersController"},"delete":{"tags":["Users"],"parameters":[{"name":"id","in":"path","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{}}}}},"x-samchon-accessor":["users","deleteUser"],"x-samchon-controller":"UsersController"},"get":{"tags":["Users"],"parameters":[{"name":"id","in":"path","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"x-samchon-accessor":["users","findOneUser"],"x-samchon-controller":"UsersController"}}},"components":{"schemas":{"HealthCheckResult":{"type":"object","properties":{"status":{"oneOf":[{"const":"error"},{"const":"ok"},{"const":"shutting_down"}],"description":"The overall status of the Health Check"},"info":{"$ref":"#/components/schemas/HealthIndicatorResult","description":"The info object contains information of each health indicator\nwhich is of status \"up\""},"error":{"$ref":"#/components/schemas/HealthIndicatorResult","description":"The error object contains information of each health indicator\nwhich is of status \"down\""},"details":{"$ref":"#/components/schemas/HealthIndicatorResult","title":"The details object contains information of every health indicator","description":"The details object contains information of every health indicator."}},"required":["status","details"],"description":"The result of a health check"},"HealthIndicatorResult":{"type":"object","properties":{},"required":[],"description":"The result object of a health indicator","additionalProperties":{"type":"object","properties":{"status":{"oneOf":[{"const":"up"},{"const":"down"}],"description":"The status if the given health indicator was successful or not"}},"required":["status"],"additionalProperties":{}}},"FileDto":{"type":"object","properties":{"originalName":{"type":"string","description":"원본 파일명"},"mimeType":{"type":"string","description":"파일 MIME 타입"},"size":{"type":"number","description":"파일 크기 (bytes)"},"storageType":{"oneOf":[{"const":"local"},{"const":"s3"}],"description":"저장소 타입"},"metadata":{"type":"string","description":"파일 메타데이터"},"id":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}},"required":["originalName","mimeType","size","storageType","metadata","id","createdAt","updatedAt"]},"UserRegisterDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"}},"required":["firstName","lastName","email"]},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}},"required":["id","firstName","lastName","email","createdAt","updatedAt"]}}},"tags":[{"name":"helloworld"},{"name":"Files"},{"name":"Users"}],"x-samchon-emended":true}